feat: get single user by id - route,  added fetch profile payload
